重构
重构了整体，支持了react
