+++ src/index.ts
@@ -1,118 +1,17 @@
-import {
-  ResolvedConfig,
-  type Plugin,
-  loadEnv,
-  normalizePath,
-  ConfigEnv,
-} from "vite";
+import { ResolvedConfig, type Plugin, ConfigEnv } from "vite";
+import { OutputOptions } from "rollup";
 import { devUpdate, devFileName } from "./devUpdate";
 import path from "node:path";
 import fs from "node:fs";
-import { load } from "cheerio";
-import {
-  type EnvSetting,
-  type TypeInput,
-  type ScriptList,
-} from "kintone-types";
+import { type TypeInput } from "kintone-types";
+import kintoneModuleInject from "./kintoneModuleInject";
+import getServerInfo from "./getServerInfo";
+import getIndexScripts from "./getIndexScripts";
+import getEnvInfo from "./getEnvInfo";
+import getEntry from "./getEntry";
+import getDirFiles from "./getDirFiles";
 
-//类型守卫函数
-function isEnvSetting(obj: any): obj is EnvSetting {
-  return (
-    obj &&
-    typeof obj.VITE_KINTONE_URL === "string" &&
-    typeof obj.VITE_KINTONE_USER_NAME === "string" &&
-    typeof obj.VITE_KINTONE_PASSWORD === "string" &&
-    (typeof obj.VITE_KINTONE_APP === "undefined" ||
-      typeof obj.VITE_KINTONE_APP === "string")
-  );
-}
-
-function getIndexHtmlContent(): ScriptList {
-  const url = path.resolve("index.html");
-  const htmlContent = fs.readFileSync(url, "utf-8");
-
-  const $ = load(htmlContent);
-  const scriptTags = $("body script");
-
-  const scriptList: ScriptList = [];
-  scriptTags.each((index, element) => {
-    const data = {
-      type: $(element).attr("type"),
-      src: $(element).attr("src"),
-    };
-    scriptList.push(data);
-  });
-  return scriptList;
-}
-
-function kintoneModuleHack(
-  devServerUrl: string,
-  scriptList: ScriptList
-): string {
-  return `(function () {
-    const viteClientInject = document.createElement("script");
-    viteClientInject.type = "module";
-    viteClientInject.src = "${devServerUrl}"+'/@vite/client';
-    document.body.appendChild(viteClientInject);
-    const scriptElement = document.createElement("script");
-    scriptElement.type = "module";
-    scriptElement.textContent = \`import RefreshRuntime from '${devServerUrl}/@react-refresh';
-    RefreshRuntime.injectIntoGlobalHook(window);
-    window.$RefreshReg$ = () => {};
-    window.$RefreshSig$ = () => (type) => type;
-    window.__vite_plugin_react_preamble_installed__ = true;\`;
-    document.body.appendChild(scriptElement);
-    const scriptList = ${JSON.stringify(scriptList)};
-    function loadScript(src,type) {
-      const script = document.createElement("script");
-      script.type = type;
-      script.src = "${devServerUrl}"+src;
-      document.body.appendChild(script);
-    }
-    for (const script of scriptList){
-      const {src,type}=script
-      loadScript(src,type)
-    }
-  })();
-  `;
-}
-
-function getDirFiles(dir: string, extList: string[]) {
-  let result: string[] = [];
-  let files = fs.readdirSync(dir, { withFileTypes: true });
-  files.forEach((file) => {
-    const filepath = path.join(dir, file.name);
-    const ext = path.extname(filepath).slice(1);
-    if (file.isFile() && extList.includes(ext)) {
-      result.push(filepath);
-    } else if (file.isDirectory()) {
-      result.push(...getDirFiles(filepath, extList));
-    }
-  });
-  return result;
-}
-
-function validateEnv(
-  envConfig: ConfigEnv,
-  viteConfig: ResolvedConfig
-): EnvSetting | undefined {
-  const resolvedRoot = normalizePath(
-    viteConfig.root ? path.resolve(viteConfig.root) : process.cwd()
-  );
-
-  const envDir = viteConfig.envDir
-    ? normalizePath(path.resolve(resolvedRoot, viteConfig.envDir))
-    : resolvedRoot;
-
-  const env = loadEnv(envConfig.mode, envDir, viteConfig.envPrefix);
-  return isEnvSetting(env) ? env : undefined;
-}
-
-function isIpv6(address: any) {
-  return address.family === "IPv6" || address.family === 6;
-}
-
-export default function kintoneDev(inputType: TypeInput): Plugin[] {
+export default function kintoneDev(options: TypeInput): Plugin[] {
   let viteConfig: ResolvedConfig;
   let envConfig: ConfigEnv;
 
@@ -120,47 +19,36 @@ export default function kintoneDev(inputType: TypeInput): Plugin[] {
     {
       name: "vite-plugin-kintone-dev:dev",
       apply: "serve",
-      enforce: "post", // 指定运行顺序
+      enforce: "post",
       config: (config, env) => (envConfig = env),
       configResolved(config) {
         viteConfig = config;
-        console.log(viteConfig.server);
-        viteConfig.server.origin = "http://127.0.0.1:8080";
       },
       configureServer(server) {
-        // console.log("server", server);
         server.httpServer?.once("listening", async () => {
-          const outputDir = path.resolve(viteConfig.build.outDir);
-          const address = server.httpServer?.address();
-          if (!address || typeof address === "string") {
-            console.error("Unexpected dev server address", address);
-            process.exit(1);
+          const devServerUrl = getServerInfo(server);
+          if (!server.config.server.origin) {
+            server.config.server.origin = devServerUrl;
           }
-          const protocol = server.config.server.https ? "https" : "http";
-          const host = isIpv6(address)
-            ? `[${address.address}]`
-            : address.address;
-          const port = address.port;
-          const devServerUrl = `${protocol}://${host}:${port}`;
-          // server.origin = devServerUrl;
+          const outputDir = path.resolve(viteConfig.build.outDir);
           if (!fs.existsSync(outputDir)) {
             fs.mkdirSync(outputDir);
           }
+          const scriptList = getIndexScripts();
           const fileUrl = path.resolve(outputDir, devFileName);
-
-          const scriptList = getIndexHtmlContent();
-
           fs.writeFileSync(
             fileUrl,
-            kintoneModuleHack(devServerUrl, scriptList)
+            kintoneModuleInject(devServerUrl, scriptList, options.react)
           );
-
-          const env = validateEnv(envConfig, viteConfig);
-
+          const env = getEnvInfo(envConfig, viteConfig);
           if (env) {
-            devUpdate(env, [fileUrl], inputType).then((r) => {
-              fs.unlinkSync(fileUrl);
-            });
+            devUpdate(env, [fileUrl], options)
+              .then((r) => {
+                fs.unlinkSync(fileUrl);
+              })
+              .catch(() => {
+                console.log("upload failed");
+              });
           } else {
             console.log("env error");
           }
@@ -172,16 +60,36 @@ export default function kintoneDev(inputType: TypeInput): Plugin[] {
       apply: "build",
       enforce: "post",
       config(config, env) {
+        const entry = getEntry(config);
         envConfig = env;
+
         config.build = {
+          modulePreload: { polyfill: false },
+          manifest: true,
           cssCodeSplit: false,
           rollupOptions: {
+            input: entry,
             output: {
               format: "iife",
             },
           },
         };
+
+        if (options?.build?.outputName !== undefined) {
+          (
+            config.build.rollupOptions?.output as OutputOptions
+          ).entryFileNames = `${options?.build?.outputName}.js`;
+          (config.build.rollupOptions?.output as OutputOptions).assetFileNames =
+            (assetInfo): string => {
+              if (assetInfo.name?.endsWith(".css")) {
+                return `${options?.build?.outputName}[extname]`;
+              } else {
+                return assetInfo.name as string;
+              }
+            };
+        }
       },
+
       configResolved(config) {
         viteConfig = config;
       },
@@ -189,9 +97,14 @@ export default function kintoneDev(inputType: TypeInput): Plugin[] {
         const outputDir = path.resolve(viteConfig.build.outDir);
         const extList = ["js", "css"];
         const fileList = getDirFiles(outputDir, extList);
-        const env = validateEnv(envConfig, viteConfig);
+        const env = getEnvInfo(envConfig, viteConfig);
         if (env) {
-          devUpdate(env, fileList, inputType);
+          //是否需要根据output name来判断是否进行上传？
+          if (options.build?.upload) {
+            devUpdate(env, fileList, options).catch(() => {
+              console.log("upload failed");
+            });
+          }
         } else {
           console.log("env error");
         }
